STM32 Chip update
For some reason, for which I don't know, it is impossible to use tri state logic consistently with the STM32G431KBT6. There is, im guessing, parasitic capacitance between Port A and Port B pins that prevent a pin from ever being specified when the other port (same pin #) is high, low, or Z. I tested this out as well with the NUCLEO-G431KB, the commercial dev board for the STM32G431KBT6. I will be focusing on using an ATMEGA 4808 chip instead.