Update - V3.0
I lied. I said I was going to stop working on hardware and only focus on software. But something was bothering me. The two mistakes in version 2.0 were: one of the pins was always high (no idea) and two, the charge sense pin was NOT on an asynchronous pin - meaning the watch wouldn't know it was charging (it was charging, but it didn't know), until a button was pressed. Now, in V3.0, the the charge sense is routed to an asynchronous pin, so when it's plugged in it can show can indicator automatically (with no button press)! I also got rid of the nuts on the back to screw in the backplate and went for a compression fitting - it looks a lot cleaner and more components can be accessed. Also - its now in black and 1 mm thinner (10% thinner)!