Near Final Update
Found out the BOD controller only shuts down the chip control instead of shutting down the functionality of the pins. I have rewrote the software to use a different control function (VLM instead of BOD) so it should (in theory) be able to shut down safely.

Additionally, I re-wired 2 traces on the circuit board so that the battery-percentage module is 100% safe and will not drain the battery. It's not worth paying the money to get new boards printed, since there's only one non-critical wire added to the back of the board. If the wire breaks contact, the battery read module will read 99% since there's no "ground" connected.

This needs to be tested, of course.